# swagger.yaml
swagger: "2.0"
info:
  description: "Movie app rest api service."
  version: "1.0.0"
  title: "Swagger MovieApp RestApi"
  contact:
    email: "classlifedavis@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/api/v1"
tags:
- name: "movie"
  description: "Endpoints for Movie model"
schemes:
- "https"
- "http"
paths:
  /movies:
    get:
      tags:
      - "movie"
      summary: "get the list of movies"
      description: ""
      operationId: "getMovie"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "201":
          description: "Success"
      security:
      - movieapp_auth:
        - "read:movies"
    post:
      tags:
      - "movie"
      summary: "Add a new movie to the list"
      description: ""
      operationId: "addMovie"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Movie object that needs to be added to the list"
        required: true
        schema:
          $ref: "#/definitions/Movie"
      responses:
        "405":
          description: "Invalid input"
      security:
      - movieapp_auth:
        - "write:taks"
        - "read:movies"
  /movie/{movieId}:
    patch:
      tags:
      - "movie"
      summary: "Update an existing movie"
      description: ""
      operationId: "updateMovie"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "movieId"
        in: "path"
        description: "id of the movie that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Movie object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Movie"
      responses:
        "404":
          description: "Movie not found"
        "405":
          description: "Validation exception"
      security:
      - movieapp_auth:
        - "write:pets"
        - "read:pets"
securityDefinitions:
  movieapp_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:movies: "modify movies in your account"
      read:pets: "read your movies"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Movie:
    type: "object"
    properties:
      _id:
        type: "Object"
      title:
        type: "string"
      decription:
        type: "string"
      releaseDate:
        type: "string"
        format: "date"
      status:
        type: "string"
        default: "draft"
      labels:
        type: "array"
        items:
          type: string
    xml:
      name: "Movie"